backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

google-generativeai==0.4.1

python-dotenv==1.0.1

pypdf==4.2.0
python-multipart==0.0.9

pymongo==4.6.1
certifi
google-genai
beautifulsoup4

# LangChain for tool calling
langchain
langchain-google-genai
langchain-core
langchain-hyperbrowser

# Authentication
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
itsdangerous==2.1.2
google-auth==2.27.0
aiosmtplib==3.0.1
bcrypt==4.1.2
